Unify and simplify legacy api installation
At the same time we get rid of the need for Store.ImportPrefix to filter the RESTMapper for each
legacy group. Before this commit we had overlapping RESTMapper, even inconsistent ones because
the root kinds were wrong outside of each legacy group.
