--- pkg/security/apis/security/install/apigroup.go
@@ -1,33 +1,0 @@
-package install
-
-import (
-	"k8s.io/apimachinery/pkg/apimachinery/announced"
-	"k8s.io/apimachinery/pkg/apimachinery/registered"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/sets"
-	kapi "k8s.io/kubernetes/pkg/api"
-
-	securityapi "github.com/openshift/origin/pkg/security/apis/security"
-	securityapiv1 "github.com/openshift/origin/pkg/security/apis/security/v1"
-)
-
-func installApiGroup() {
-	Install(kapi.GroupFactoryRegistry, kapi.Registry, kapi.Scheme)
-}
-
-// Install registers the API group and adds types to a scheme
-func Install(groupFactoryRegistry announced.APIGroupFactoryRegistry, registry *registered.APIRegistrationManager, scheme *runtime.Scheme) {
-	if err := announced.NewGroupMetaFactory(
-		&announced.GroupMetaFactoryArgs{
-			GroupName:                  securityapi.GroupName,
-			VersionPreferenceOrder:     []string{securityapiv1.SchemeGroupVersion.Version},
-			RootScopedKinds:            sets.NewString("SecurityContextConstraints"),
-			AddInternalObjectsToScheme: securityapi.AddToScheme,
-		},
-		announced.VersionToSchemeFunc{
-			securityapiv1.SchemeGroupVersion.Version: securityapiv1.AddToScheme,
-		},
-	).Announce(groupFactoryRegistry).RegisterAndEnable(registry, scheme); err != nil {
-		panic(err)
-	}
-}
